scripts/series_gen/src/requirements.txt
pyyaml==5.3.1
python-slugify